+++ src/App.js
@@ -1,11 +1,11 @@
-import React from 'react';
-import './App.css';
-import SimpleFileUploader from "./components/FileUploader";
+import React from "react";
+import "./App.css";
+import SimpleFileUploader from "./components/FileUploader/FileUploader";
 
 function App() {
   return (
     <div className="App">
-      <SimpleFileUploader/>
+      <SimpleFileUploader />
     </div>
   );
 }
